+++ CHANGELOG.md
@@ -8,7 +8,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
-- Refactored movement usecase findByPeriod[PR#84](https://github.com/silvioubaldino/personal-finance/pull/84)
+- Refactored movement handler findByPeriod [PR#85](https://github.com/silvioubaldino/personal-finance/pull/85)
+- Refactored movement usecase findByPeriod [PR#84](https://github.com/silvioubaldino/personal-finance/pull/84)
 - Migrated recurrent repository find queries [PR#83](https://github.com/silvioubaldino/personal-finance/pull/83)
 - Updated libs [PR#82](https://github.com/silvioubaldino/personal-finance/pull/82)
 - Refactored movement repository find queries [PR#81](https://github.com/silvioubaldino/personal-finance/pull/81)